Just compute configs, don't run esphome; fix tmpl
The main change was from trying to execute esphome and do the
programming, which was low-value, to just computing the templated
configs.  I can call esphome from the CLI just fine.  :)

I also had to explicitly turn on the I2C power early, to avoid the SDL
line behind held low by the devices that weren't on.  This wasn't
necessary before in my prototype, but it seems like a race condition in
the esphome framework means what used to work by luck now requires a bit
of care to be reliable.

This updates the esphome framework from arduino to esp-idf to mitigate
the deprecation of the legacy RMT driver that was causing a compile time
warning.  That also necessitated a change from neopixelbus to
esp32_rmt_led_strip.

I also backed down the polling interval on the CO2 sensor from 5s to 30s
for a more production-reliable rate, so I'm not overheating the sensors.